Fixes issue #2430 - Add Jacoco check to mime-type extension
