+++ extension/lite/src/main/java/module-info.java
@@ -35,11 +35,11 @@
  * </p>
  * <ul>
  *  <li>Manorrock Herring (JNDI)</li>
- *  <li>Servlet Container Initializer</li>
+ *  <li>Mime-type</li>
+ *  <li>ServletContainerInitializer</li>
  *  <li>Servlet Security</li>
  *  <li>Annotation Scanning</li>
  *  <li>Locale Encoding</li>
- *  <li>Mime-type</li>
  *  <li>Java Policy</li>
  *  <li>Servlet Annotations</li>
  *  <li>TEMPDIR</li>
@@ -51,12 +51,12 @@ module cloud.piranha.extension.lite {
     opens cloud.piranha.extension.lite;
     requires cloud.piranha.core.api;
     requires transitive cloud.piranha.extension.herring;
+    requires transitive cloud.piranha.extension.mimetype;
     requires transitive cloud.piranha.extension.scinitializer;
     requires transitive cloud.piranha.extension.security.servlet;
     requires transitive cloud.piranha.extension.servletannotations;
     requires transitive cloud.piranha.extension.standard.annotationscan;
     requires transitive cloud.piranha.extension.standard.localeencoding;
-    requires transitive cloud.piranha.extension.standard.mimetype;
     requires transitive cloud.piranha.extension.standard.policy;
     requires transitive cloud.piranha.extension.tempdir;
     requires transitive cloud.piranha.extension.webxml;
