+++ extension/pom.xml
@@ -38,7 +38,6 @@
         <module>standard-async</module>
         <module>standard-localeencoding</module>
         <module>standard-logging</module>
-        <module>standard-mimetype</module>
         <module>standard-policy</module>
         <module>tempdir</module>
         <module>wasp</module>
@@ -46,6 +45,7 @@
         <module>weld-servlet</module>
         <module>webxml</module>
         <module>welcomefile</module>
+        <module>mimetype</module>
     </modules>
 
     <dependencyManagement>
