+++ extension/webxml/src/test/java/cloud/piranha/extension/webxml/MimeMappingTest.java
@@ -28,7 +28,7 @@
 package cloud.piranha.extension.webxml;
 
 import cloud.piranha.core.impl.DefaultWebApplication;
-import cloud.piranha.extension.standard.mimetype.StandardMimeTypeManager;
+import cloud.piranha.extension.mimetype.internal.InternalMimeTypeManager;
 import cloud.piranha.resource.impl.DirectoryResource;
 import java.io.File;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -49,7 +49,7 @@ class MimeMappingTest {
     @Test
     void testGetMimeType() throws Exception {
         DefaultWebApplication webApplication = new DefaultWebApplication();
-        StandardMimeTypeManager manager = new StandardMimeTypeManager();
+        InternalMimeTypeManager manager = new InternalMimeTypeManager();
         webApplication.getManager().setMimeTypeManager(manager);
         webApplication.addResource(new DirectoryResource(new File("src/test/webxml/mimeMapping")));
         webApplication.addInitializer(new WebXmlInitializer());
